현재까지 작업 내용 추가

+++ models/ai_product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import ForeignKey, Column, Integer, Numeric
+from sqlalchemy import ForeignKey, Column, Integer, Numeric, BIGINT
 from sqlalchemy.orm import relationship
 
 from db.database import base
@@ -8,7 +8,7 @@ class AiProduct(base):
     __tablename__ = "ai_product"
     ai_product_id= Column(Integer, primary_key=True)
     ai_keyword_id = Column(Integer, ForeignKey("ai_keyword.ai_keyword_id"), nullable=False)
-    product_id = Column(Integer, ForeignKey("product.product_id"), nullable=False)
+    product_id = Column(BIGINT, ForeignKey("product.product_id"), nullable=False)
     rank = Column(Numeric(10,2), nullable=False)
 
     ai_keyword = relationship("AiKeyword", back_populates="ai_product")
